Download manager, настройки

+++ printmap-gui/src/main/java/ru/mapolib/printmap/gui/utils/FileSizeUtil.kt
@@ -1,6 +1,6 @@
 package ru.mapolib.printmap.gui.utils
 
-internal fun formatSize(bytes: Int): String {
+internal fun formatSize(bytes: Long): String {
     val units = arrayOf("Б", "КБ", "МБ", "ГБ", "ТБ")
     var size = bytes.toDouble()
     var unitIndex = 0
